refactor(BA-6672): move remaining value types down to data layer
Make the manager data layer a pure leaf by removing its residual
imports from models:

- vfolder/dto.py: import VFolder{OperationStatus,OwnershipType,
  MountPermission} directly from data.vfolder.types
- image/types.py: move the `Resources` type alias down from
  models/image/row.py; models now re-references it from data
- scaling_group/types.py: move FairShareScalingGroupSpec down from
  models/scaling_group/types.py (now removed), following the existing
  data/session and data/deployment BackendAISchema precedent
