+++ src/ai/backend/manager/models/image/__init__.py
@@ -1,12 +1,16 @@
 from ai.backend.common.types import ImageAlias
-from ai.backend.manager.data.image.types import ImageIdentifier, ImageStatus, ImageType
+from ai.backend.manager.data.image.types import (
+    ImageIdentifier,
+    ImageStatus,
+    ImageType,
+    Resources,
+)
 
 from .row import (
     ImageAliasRow,
     ImageLoadFilter,
     ImageRow,
     PublicImageLoadFilter,
-    Resources,
     bulk_get_image_configs,
 )
 from .row import (
