+++ src/ai/backend/manager/models/scaling_group/__init__.py
@@ -1,3 +1,5 @@
+from ai.backend.manager.data.scaling_group.types import FairShareScalingGroupSpec
+
 from .row import (
     ScalingGroupForDomainRow,
     ScalingGroupForKeypairsRow,
@@ -14,7 +16,6 @@ from .row import (
     sgroups_for_groups,
     sgroups_for_keypairs,
 )
-from .types import FairShareScalingGroupSpec
 
 __all__ = (
     "FairShareScalingGroupSpec",
